Moved MoreStrings.headerTokenCompatible() to HTTP.headerTokenCompatible()
